+++ build.rs
@@ -42,9 +42,9 @@ fn main() {
         // declaration they belong to.
         .generate_comments(false)
         .header(input.to_string_lossy())
-        // Make cargo invalidate the generated bindings whenever any of the files included from the
-        // header change.
-        .parse_callbacks(Box::new(bindgen::CargoCallbacks::new()))
+        // Activate parse callbacks. This causes cargo to invalidate the generated bindings when any
+        // of the included files change. It also enables us to rename items in the final bindings.
+        .parse_callbacks(Box::new(CustomCallbacks))
         // We may use `core` instead of `std`. This might be useful for `no_std` environments.
         .use_core()
         // Wrap static functions. These are used in several places for inline helpers and we want to
@@ -75,4 +75,32 @@ fn main() {
         .flag_if_supported("-Wno-deprecated-declarations")
         .flag_if_supported("-Wno-deprecated")
         .compile(ext_name);
+}
+
+#[derive(Debug)]
+struct CustomCallbacks;
+
+// Include `cargo:rerun-if` instructions just like `bindgen::CargoCallbacks` does. In addition, make
+// necessary adjustments to the names of items for the final bindings.
+impl bindgen::callbacks::ParseCallbacks for CustomCallbacks {
+    fn header_file(&self, filename: &str) {
+        // Make sure to rerun build when header file changes.
+        println!("cargo:rerun-if-changed={}", filename);
+    }
+
+    fn include_file(&self, filename: &str) {
+        // Make sure to rerun build when include file changes.
+        println!("cargo:rerun-if-changed={}", filename);
+    }
+
+    fn read_env_var(&self, key: &str) {
+        // Make sure to rerun build when environment variable changes.
+        println!("cargo:rerun-if-env-changed={}", key);
+    }
+
+    fn item_name(&self, original_item_name: &str) -> Option<String> {
+        // Rename pointer constant back to its original name. See `wrapper.c` for details.
+        (original_item_name == "UA_EMPTY_ARRAY_SENTINEL_")
+            .then(|| "UA_EMPTY_ARRAY_SENTINEL".to_owned())
+    }
 }